+++ src/pages/about-us/AboutUs.tsx
@@ -2,7 +2,8 @@ import { Box, Grid, Link, Paper, Typography } from '@mui/material';
 import { motion } from 'motion/react';
 import { Link as ReactLink } from 'react-router';
 
-import { Title } from '../../components';
+import Title from '@/components/title/Title';
+
 import FlipCard from './components/flip-card/FlipCard';
 import { COLLABORATION, DEVELOPERS, MOTTOS } from './data';
 import rsLogo from './imgs/rss-logo.c19ce1b4.svg';
