+++ src/pages/profile/Profile.tsx
@@ -1,10 +1,12 @@
 import { Box, Button, Chip, Stack } from '@mui/material';
 import { motion } from 'motion/react';
 
-import { Loading, Title } from '../../components';
-import { useAppDispatch, useAppSelector } from '../../redux/hooks';
-import { getCustomer } from '../../redux/selectors';
-import { setIsEditProfile } from '../../redux/slices/authSlice';
+import Loading from '@/components/loading/Loading';
+import Title from '@/components/title/Title';
+import { useAppDispatch, useAppSelector } from '@/redux/hooks';
+import { getCustomer } from '@/redux/selectors';
+import { setIsEditProfile } from '@/redux/slices/authSlice';
+
 import AuthLayout from './auth-layout/AuthLayout';
 import GuestLayout from './guest-layout/GuestLayout';
