+++ vite.config.ts
@@ -1,13 +1,24 @@
 /// <reference types="vitest" />
 import netlifyPlugin from '@netlify/vite-plugin-react-router';
 import react from '@vitejs/plugin-react';
+import path from 'path';
 import { defineConfig } from 'vite';
 import eslint from 'vite-plugin-eslint2';
 
 // https://vite.dev/config/
 export default defineConfig({
   plugins: [react(), eslint(), netlifyPlugin()],
-
+  resolve: {
+    alias: {
+      '@': path.resolve(__dirname, './src'),
+      '@assets': path.resolve(__dirname, './src/assets'),
+      '@components': path.resolve(__dirname, './src/components'),
+      '@pages': path.resolve(__dirname, './src/pages'),
+      '@ecommerce': path.resolve(__dirname, './src/ecommerce'),
+      '@redux': path.resolve(__dirname, './src/redux'),
+      '@services': path.resolve(__dirname, './src/services'),
+    },
+  },
   test: {
     globals: true,
     environment: 'jsdom',
